% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpolar.R
\docType{package}
\name{ggpolar}
\alias{ggpolar}
\title{A ggplot2 Extension to Plot Data in Polar Coordinates on a Cartesian Layer}
\description{
This package supplies a few ggplot2 facilities to plot data in polar coordinates
 on a cartesian layer. For more complete documentation, read the vignettes.
}
