% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{polar_y}
\alias{polar_y}
\title{pole_y}
\usage{
polar_y(radius, theta)
}
\arguments{
\item{radius}{a numeric}

\item{theta}{a numeric}
}
\value{
y
}
\description{
pole_y
}
\examples{
see ?stat_polar for more details.
}
