% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-polar.R
\docType{data}
\name{stat_polar}
\alias{stat_polar}
\alias{StatPolar}
\title{Plot data in polar coordinates on a cartesian layer}
\usage{
stat_polar(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Data in polar coordinates can be drawn directly on this layer through two new aesthetics:
p_theta and p_radius. Data can also be translated and/or rotated through another three
new aesthetics: translate_x, translate_y and rotate.
}
\section{Details}{

For the aesthetics p_theta and rotate, angles are in degrees, not radians.
For example, east is 0 and north is 90.
}

\section{Aesthetics}{

stat_polar() understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{p_theta}
  \item \strong{p_radius}
  \item translate_x
  \item translate_y
  \item rotate
  \item group
  }
Learn more about setting these aesthetics in vignette("ggplot2-specs").
}

\examples{
library(ggplot2)

polar_data <- data.frame(theta = c(0,120,240), radius = 1)
ggplot(polar_data) + stat_polar(aes(p_theta = theta, p_radius = radius))

# I strongly recommend that use StatPolar rather than stat_polar().
# Because, with StatPolar, you can draw data in polar coordinates
# on the layers in ggplot2 and its extension packages only in two steps:
#   1.set stat to StatPolar (stat = "polar").
#   2.map the theta and radius to two new aesthetics: p_theta and p_radius.

ggplot(polar_data, aes(p_theta = theta, p_radius = radius)) +
  geom_polygon(stat = "polar") +
  geom_path(stat = "polar", colour = "green", size = 3) +
  geom_point(stat = "polar", colour = "red", size = 6)

# You can also translate and/or rotate the data through translate_x, translate_y
# and rotate.
ggplot(polar_data, aes(p_theta = theta, p_radius = radius)) +
  geom_polygon(aes(colour = "original"), stat = "polar", fill = NA) +
  geom_polygon(aes(colour = "rotate", rotate = 60),
               stat = "polar", fill = NA) +
  geom_polygon(aes(colour = "translate",
               translate_x = 2, translate_y = 2),
               stat = "polar", fill = NA) +
  geom_polygon(aes(colour = "translate + rotate",
               translate_x = 2, translate_y = 2, rotate = 60),
               stat = "polar", fill = NA) +
  scale_color_discrete(name = NULL) +
  scale_size(guide = NULL) +
  coord_equal()

# ggpolar also can translate and rotate data in cartesian coordinates
# by providing 2 helper functions, xy_theta() and xy_radius():

# cartesian coordinates
cartiesian_data <- data.frame(x = c(1,-1,-1,1), y = c(1,1,-1,-1))

ggplot(cartiesian_data, aes(p_theta = xy_theta(x,y),
                            p_radius = xy_radius(x,y))) +
 geom_polygon(aes(fill = "original"), stat = "polar") +
 geom_polygon(aes(rotate = 45, translate_x = 2, translate_y = 2),
              stat = "polar") +
 scale_fill_discrete(name = NULL)

# For more examples, see the vignette 'The complete guide for ggpolar'.
}
\keyword{datasets}
