% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-polar.R
\docType{data}
\name{StatPolar}
\alias{StatPolar}
\title{StatPolar}
\format{An object of class \code{StatPolar} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatPolar
}
\description{
StatPolar
}
\section{Aesthetics}{

new aes: radius, theta, x0, y0
}

\keyword{datasets}
