% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-pole.R
\docType{data}
\name{StatPole}
\alias{StatPole}
\title{StatPole}
\format{An object of class \code{StatPole} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatPole
}
\description{
StatPole
}
\keyword{datasets}
